pip_requirements.txt
absl-py==0.12.0
awscli==1.19.94
backcall==0.2.0
botocore==1.20.94
certifi==2021.5.30
colorama==0.4.3
coverage==5.5
cycler==0.10.0
Cython==0.29.23
decorator==5.0.9
docutils==0.15.2
easydict==1.9
grpcio==1.36.1
importlib-metadata==3.10.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.18.0
jmespath==0.10.0
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
mkl-fft==1.3.0
mkl-random==1.2.1
mkl-service==2.3.0
numpy==1.20.2
olefile==0.46
pandas==1.1.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pip==21.1.2
prompt-toolkit==3.0.17
protobuf==3.14.0
ptyprocess==0.7.0
pyasn1==0.4.8
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==20.0.0
rsa==4.7.2
s3transfer==0.4.2
scipy==1.6.2
setuptools==52.0.0.post20210125
sip==4.19.13
six==1.16.0
tensorboard==1.15.0
torch==1.8.1
torchaudio==0.8.0a0+e4e171a
tornado==6.1
tqdm==4.61.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.5
wcwidth==0.2.5
Werkzeug==1.0.1
wheel==0.36.2
zipp==3.4.1
